+++ tests/testthat/test-hdb_query.R
@@ -15,3 +15,11 @@ test_that("verify_dates works as expected", {
     vd("2020-1-1", "2020-12-17", "MN")
   )
 })
+
+test_that("query fails when end_date is before start_date", {
+  expect_error(hdb_query(21327, "lc", "m", "2020-09", "2020-08", 3125))
+
+  expect_error(hdb_query(1840, "uc", "d", "2020-09-01", "2020-08-01"))
+
+  expect_error(hdb_query(1792, "uc", "h", "2020-09-01 9:00", "2020-09-01 6:00"))
+})